refactor: remove axios and move get distance to helpers
