+++ back-end/src/helpers/getCoordinates.ts
@@ -1,12 +1,11 @@
-import axios from "axios";
 import { CustomError } from "../errors/customError";
 
 export default async function getCoordinates(zipcode: string) {
   const url = `https://nominatim.openstreetmap.org/search?postalcode=${zipcode}&format=json`;
 
   try {
-    const response = await axios.get(url);
-    const data = response.data;
+    const response = await fetch(url);
+    const data = await response.json();
 
     if (data && data.length > 0) {
       return {
