+++ back-end/src/services/eventService.ts
@@ -4,8 +4,8 @@ import { Location } from "../models/location";
 import { Category } from "../models/category";
 import { CustomError } from "../errors/customError";
 import { City } from "../models/city";
-import axios from "axios";
 import { Op } from "sequelize";
+import getDistance from "../helpers/getDistance";
 
 type EventData = Event &
   Partial<Location> & {
@@ -45,6 +45,7 @@ class EventService {
 
     return events;
   }
+
   static async getEventsByDate(
     cityId: string | undefined,
     cityName: string | undefined
@@ -130,11 +131,10 @@ class EventService {
         latitude: event.Location.coordlat,
         longitude: event.Location.coordlon,
       };
-      const distance = EventService.getDistance(
+      const distance = getDistance(
         userCoordinates,
         eventCoordinates
       );
-
       let recommendationPoints = 0;
 
       if (distance <= 1.0) {
@@ -152,17 +152,8 @@ class EventService {
           )
         ) {
           recommendationPoints += 5;
-          console.log(`\n\nUsuário gosta de ${userCategory.name} +5 pontos`);
         }
       });
-
-      console.log(
-        `A distância entre o usuário e o evento ${event.name} é de ${distance} km.`
-      );
-
-      console.log(
-        `O evento ${event.name} recebeu ${recommendationPoints} pontos.\n\n`
-      );
 
       return {
         ...event.get({ plain: true }),
@@ -176,33 +167,6 @@ class EventService {
     );
 
     return resolvedEvents;
-  }
-
-  static getDistance(coord1: any, coord2: any) {
-    const lat1 = coord1.latitude;
-    const lon1 = coord1.longitude;
-    const lat2 = coord2.latitude;
-    const lon2 = coord2.longitude;
-
-    if (lat1 == lat2 && lon1 == lon2) {
-      return 0;
-    } else {
-      var radlat1 = (Math.PI * lat1) / 180;
-      var radlat2 = (Math.PI * lat2) / 180;
-      var theta = lon1 - lon2;
-      var radtheta = (Math.PI * theta) / 180;
-      var dist =
-        Math.sin(radlat1) * Math.sin(radlat2) +
-        Math.cos(radlat1) * Math.cos(radlat2) * Math.cos(radtheta);
-      if (dist > 1) {
-        dist = 1;
-      }
-      dist = Math.acos(dist);
-      dist = (dist * 180) / Math.PI;
-      dist = dist * 60 * 1.1515;
-      dist = dist * 1.609344;
-      return dist;
-    }
   }
 
   static async createEvent(eventData: EventData): Promise<Event> {
